Write most of the first draft of CONTRIBUTING.md
This document serves as a working standard for how to use git
professionally in the CCTC repository. This document should serve as
a starting point for any future GitLab version control standards used
by the schoolhouse.

A lot of external resources are referenced in order to provide more
in-depth reading on various topics. These should be discussed for
their merit and whether, for example, a rebase-based approach is
better than a merge-based approach to handling changes and conflicts
during the standard git feature branch workflow process.