Finish first draft of CONTRIBUTING.md
This is my initial draft of how I envision the schoolhouse should use GitLab and
version control in general. This document is tailored specifically to the CCTC
repository.

Some future improvements could be:
  - Code style standards
  - Code of conduct (aka "no asshole policy")
  - Merge request standards

This document should always be seen as a working document. It
should also be mandatory reading for any new instructors or
contributers to the repository.