Modify net1.sh and nix1_net3.sh scripts
apt-get upgrade has been moved from after the pkg_array installation loop
to before this because it just makes more sense.

Tinkering with the pkg_array install process. Fixed the main issue where
the presense of linux-headers-($uname -r) was breaking things. I'm pretty
sure whoever wrote it meant for it to be $(uname -r), but I'm just
removing it entirely for now because I'm pretty sure its not necessary
to update the kernel headers.

Once the basic installation of services and packages works, I'll mess
with the telnetd service and have it use systemctl instead of xinetd.