Revert master yaml configuration
I had to change the nix1_net3 configuration so that it uses the
template in my branch rather than the template in the master branch.
This commit simply reverts the template back to the master script.

Once this branch is merged to master, it will then obviously point
to the fixed shell script.